feat(help packages): add token placer as we do for html
